requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
black==20.8b1
click==7.1.2
dataclasses==0.8
Django==3.1.6
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
flake8==3.8.4
flake8-django==1.1.1
importlib-metadata==3.4.0
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.2
pytest-black==0.3.12
pytz==2021.1
regex==2020.11.13
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
zipp==3.4.0
